+++ test/utils.js
@@ -34,12 +34,12 @@ describe('utils.parseHostString() #noserver', function () {
 
   it('parses a domain name with TLS name and port', function () {
     var host = utils.parseHostString('aero.local:aero.tls:3333')
-    expect(host).to.eql({ addr: 'aero.local', tls: 'aero.tls', port: 3333 })
+    expect(host).to.eql({ addr: 'aero.local', tlsname: 'aero.tls', port: 3333 })
   })
 
   it('parses a domain name with TLS name', function () {
     var host = utils.parseHostString('aero.local:aero.tls')
-    expect(host).to.eql({ addr: 'aero.local', tls: 'aero.tls', port: 3000 })
+    expect(host).to.eql({ addr: 'aero.local', tlsname: 'aero.tls', port: 3000 })
   })
 
   it('parses an IPv4 address', function () {
@@ -54,12 +54,12 @@ describe('utils.parseHostString() #noserver', function () {
 
   it('parses an IPv4 address with TLS name and port', function () {
     var host = utils.parseHostString('192.168.33.10:aero.tls:3333')
-    expect(host).to.eql({ addr: '192.168.33.10', tls: 'aero.tls', port: 3333 })
+    expect(host).to.eql({ addr: '192.168.33.10', tlsname: 'aero.tls', port: 3333 })
   })
 
   it('parses an IPv4 address with TLS name', function () {
     var host = utils.parseHostString('192.168.33.10:aero.tls')
-    expect(host).to.eql({ addr: '192.168.33.10', tls: 'aero.tls', port: 3000 })
+    expect(host).to.eql({ addr: '192.168.33.10', tlsname: 'aero.tls', port: 3000 })
   })
 
   it('parses an IPv6 address', function () {
@@ -74,12 +74,12 @@ describe('utils.parseHostString() #noserver', function () {
 
   it('parses an IPv6 address with TLS name and port', function () {
     var host = utils.parseHostString('[fde4:8dba:82e1::c4]:aero.tls:3333')
-    expect(host).to.eql({ addr: 'fde4:8dba:82e1::c4', tls: 'aero.tls', port: 3333 })
+    expect(host).to.eql({ addr: 'fde4:8dba:82e1::c4', tlsname: 'aero.tls', port: 3333 })
   })
 
   it('parses an IPv6 address with TLS name', function () {
     var host = utils.parseHostString('[fde4:8dba:82e1::c4]:aero.tls')
-    expect(host).to.eql({ addr: 'fde4:8dba:82e1::c4', tls: 'aero.tls', port: 3000 })
+    expect(host).to.eql({ addr: 'fde4:8dba:82e1::c4', tlsname: 'aero.tls', port: 3000 })
   })
 
   it('throws an error if it cannot parse the string', function () {